task/2- Made reusable `CustomThread` class
-- `AppRunner` is moved to `Utility` package and created to maintain `DRY`
